pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "PEFT methods comparison for summarization"
authors = ["Vladislav <ve.kirienko7@gmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
peft = "^0.15.2"
pydantic = "^2.11.7"
ruff = "^0.12.0"
datasets = "^3.6.0"
trl = "^0.19.0"
ipywidgets = "^8.1.7"
python-dotenv = "^1.1.0"
wandb = "^0.20.1"
evaluate = "^0.4.4"
absl-py = "^2.3.1"
nltk = "^3.9.2"
rouge-score = "^0.1.2"
bert-score = "^0.3.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
